backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
beautifulsoup4
requests
langchain
langchain-google-genai
python-dotenv
pydantic
pypdf
python-multipart